added everything from templates/index into templates/blog
